+++ management/model_enum_certificate_key_signagure_algorithm.go
@@ -15,16 +15,14 @@ import (
 	"fmt"
 )
 
-// EnumCertificateKeySignagureAlgorithm Specifies the signature algorithm of the key. For RSA keys, options are `SHA224withRSA`, `SHA256withRSA`, `SHA384withRSA` and `SHA512withRSA`. For elliptical curve (EC) keys, options are `SHA224withECDSA`, `SHA256withECDSA`, `SHA384withECDSA`, and `SHA512withECDSA`
+// EnumCertificateKeySignagureAlgorithm Specifies the signature algorithm of the key. For RSA keys, options are `SHA256withRSA`, `SHA384withRSA` and `SHA512withRSA`. For elliptical curve (EC) keys, options are `SHA256withECDSA`, `SHA384withECDSA`, and `SHA512withECDSA`
 type EnumCertificateKeySignagureAlgorithm string
 
 // List of EnumCertificateKeySignagureAlgorithm
 const (
-	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA224WITH_RSA EnumCertificateKeySignagureAlgorithm = "SHA224withRSA"
 	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA256WITH_RSA EnumCertificateKeySignagureAlgorithm = "SHA256withRSA"
 	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA384WITH_RSA EnumCertificateKeySignagureAlgorithm = "SHA384withRSA"
 	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA512WITH_RSA EnumCertificateKeySignagureAlgorithm = "SHA512withRSA"
-	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA224WITH_ECDSA EnumCertificateKeySignagureAlgorithm = "SHA224withECDSA"
 	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA256WITH_ECDSA EnumCertificateKeySignagureAlgorithm = "SHA256withECDSA"
 	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA384WITH_ECDSA EnumCertificateKeySignagureAlgorithm = "SHA384withECDSA"
 	ENUMCERTIFICATEKEYSIGNAGUREALGORITHM_SHA512WITH_ECDSA EnumCertificateKeySignagureAlgorithm = "SHA512withECDSA"
@@ -32,11 +30,9 @@ const (
 
 // All allowed values of EnumCertificateKeySignagureAlgorithm enum
 var AllowedEnumCertificateKeySignagureAlgorithmEnumValues = []EnumCertificateKeySignagureAlgorithm{
-	"SHA224withRSA",
 	"SHA256withRSA",
 	"SHA384withRSA",
 	"SHA512withRSA",
-	"SHA224withECDSA",
 	"SHA256withECDSA",
 	"SHA384withECDSA",
 	"SHA512withECDSA",