+++ management/CHANGELOG.md
@@ -1,3 +1,17 @@
+# v0.28.0 (Unreleased)
+
+* **Breaking Change** Removed deprecated enum values `SHA224withRSA` and `SHA224withECDSA` from the `EnumCertificateKeySignagureAlgorithm` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+* **Enhancement** Added enum values to `EnumApplicationWSFEDIDPSigningAlgorithm` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+* **Enhancement** Added default value to model attributes that use the `EnumApplicationSAMLSloBinding` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+* **Enhancement** Added default value to model attributes that use the `EnumGatewayTypeLDAPSecurity` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+* **Enhancement** Added default value to model attributes that use the `EnumIdentityProvider` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+* **Enhancement** Added default value to model attributes that use the `EnumSchemaAttributeType` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+* **Enhancement** Added default value to model attributes that use the `EnumTemplateContentPushCategory` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+* **Enhancement** Added default value to the `AssertionSigned` attributes on the `ApplicationSAML` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+* **Enhancement** Added default value to the `ResponseSigned` attributes on the `ApplicationSAML` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+* **Enhancement** Added support for the `AuthnRequestSigned` attribute on the `ApplicationSAMLAllOfSpVerification` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+* **Enhancement** Added support for the `Algorithm` attribute on the `IdentityProviderSAMLAllOfSpSigning` model. [#247](https://github.com/patrickcping/pingone-go-sdk-v2/pull/247)
+
 # v0.27.0 (2023-09-05)
 
 * **Enhancement** Added `EnableRequestedAuthnContext` to the `ApplicationSAML` and `ApplicationSAMLAllOf` data models. [#245](https://github.com/patrickcping/pingone-go-sdk-v2/pull/245)